requirements.txt
boto3
python-dotenv
langchain
langchain-aws
langchain-community 
langchain-core
pypdf
faiss-cpu 
streamlit
streamlit_authenticator
awscli